Tests/LibWeb/Text/expected/css/animated-background-image-timer-stops-after-base-change.txt

active animations before base change: 1
active animations after removal: 0
